build: update to cargo-shuttle version 0.49.0
The cargo-shuttle version shipped in nixpkgs right now is several
versions out-of-date and things break when trying to use the outdated
version to deploy to shuttle.

Unfortunately, a simple override of the nixpkgs version doesn't seem to
work because `buildRustPackage` still has some annoying quirks.

After fiddling with it for a while, I decided the easiest way to deal
with this is to simply copy-paste the nix code and import the local
derivation.

Update the nixpkgs input and revert this commit once
cargo-shuttle version 0.49.0 or higher lands in nixpkgs.

Signed-off-by: Linda Siemons <linda@catbrained.dev>